ci: trigger re-run for lint check
